Feature/calculo

+++ src/main/java/com/rlti/rh/funcionario/application/request/SalarioBaseRequest.java
@@ -5,6 +5,6 @@ import java.math.BigDecimal;
 public record SalarioBaseRequest(
         Long idSalarioBase,
         BigDecimal valorSalario,
-        Long nivel
+        Integer nivel
 ) {
 }
